+++ src/apps/apps.cpp
@@ -2,6 +2,7 @@
 
 #include "launcher/registry.h"
 #include "counter/counter.h"
+#include "events/events.h"
 #include "home/home.h"
 #include "scoreboard/scoreboard.h"
 #include "settings/settings.h"
@@ -15,6 +16,7 @@ void frij_register_apps(void)
 {
     frij_registry_add(home_app());
     frij_registry_add(todo_app());
+    frij_registry_add(events_app());
     frij_registry_add(stopwatch_app());
     frij_registry_add(scoreboard_app());
     frij_registry_add(counter_app());