+++ src/apps/events/events.h
@@ -1,0 +1,16 @@
+#ifndef FRIJ_APP_EVENTS_H
+#define FRIJ_APP_EVENTS_H
+
+#include "app.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+const frij_app_t* events_app(void);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif  // FRIJ_APP_EVENTS_H